refactor: move Locale from entity to vo
